rework camdozaal fishing
